# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  backend_internal_adapter_presenter.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
      total:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    name: Amplified Access
  description: Amplified Access Watchtower — incident reporting and monitoring platform
    API.
  title: Watchtower API
  version: "1.0"
paths:
...
  /datasets:
    get:
      description: Returns a paginated list of publicly available datasets, optionally
        filtered by category
      parameters:
      - description: Page size (default 20, max 100)
        in: query
        name: limit
        type: integer
      - description: Pagination offset
        in: query
        name: offset
        type: integer
      - description: Search term
        in: query
        name: search
        type: string
      - description: 'Sort field (default: created_at)'
        in: query
        name: sort
        type: string
      - description: 'Sort direction: asc or desc'
        in: query
        name: sortOrder
        type: string
      - description: Filter by category
        in: query
        name: category
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
      summary: List public datasets
      tags:
      - Datasets
  /datasets/{id}:
    get:
      description: Returns a single public dataset by its UUID
      parameters:
      - description: Dataset UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
      summary: Get dataset by ID
      tags:
      - Datasets
  /datasets/{id}/download:
    post:
      description: Increments the download counter for a dataset
      parameters:
      - description: Dataset UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
      summary: Record dataset download
      tags:
      - Datasets
  /datasets/categories:
    get:
      description: Returns all distinct dataset categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
      summary: List dataset categories
      tags:
      - Datasets
...
  /superadmin/datasets:
    get:
      description: Returns a paginated list of all datasets regardless of visibility
        (super admin only)
      parameters:
      - description: Page size (default 20, max 100)
        in: query
        name: limit
        type: integer
      - description: Pagination offset
        in: query
        name: offset
        type: integer
      - description: Search term
        in: query
        name: search
        type: string
      - description: Sort field
        in: query
        name: sort
        type: string
      - description: 'Sort direction: asc or desc'
        in: query
        name: sortOrder
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
      security:
      - BearerAuth: []
      summary: List all datasets
      tags:
      - Super Admin – Datasets
    post:
      consumes:
      - application/json
      description: Creates a new dataset record (super admin only)
      parameters:
      - description: Dataset payload
        in: body
        name: body
        required: true
        schema:
          properties:
            category:
              type: string
            coverage:
              type: string
            description:
              type: string
            fileKey:
              type: string
            fileName:
              type: string
            fileSize:
              type: integer
            fileType:
              type: string
            format:
              type: string
            isPublic:
              type: boolean
            keywords:
              items:
                type: string
              type: array
            license:
              type: string
            methodology:
              type: string
            source:
              type: string
            tags:
              items:
                type: string
              type: array
            title:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
      security:
      - BearerAuth: []
      summary: Create dataset
      tags:
      - Super Admin – Datasets
  /superadmin/datasets/{id}:
    delete:
      description: Permanently removes a dataset (super admin only)
      parameters:
      - description: Dataset UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
      security:
      - BearerAuth: []
      summary: Delete dataset
      tags:
      - Super Admin – Datasets
...
securityDefinitions:
  BearerAuth:
    description: 'Bearer token obtained from authentication. Format: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"